micro correccion
cambio tipo de dato

+++ blog/database/migrations/2025_08_30_093836_comentarios_migration.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
        Schema::create('comentarios', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('user_id');
-            $table->unsignedBigInteger('post_id');
+            $table->Integer('user_id');
+            $table->Integer('post_id');
             $table->text('content');
             $table->timestamps();
